requirements.txt

dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-mathfilters==1.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
requests==2.25.1
psycopg2==2.8.4
django_heroku
